+++ src/Wexflow.Tasks.MediaInfo/Wexflow.Tasks.MediaInfo.csproj
@@ -59,6 +59,7 @@
   <ItemGroup>
     <Compile Include="MediaInfo.cs" />
     <Compile Include="MediaInfoLib.cs" />
+    <Compile Include="MediaInfoList.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
